Adding a window class and implemented it as a Singleton pattern.
Can acces Subject Subclasses in Observer subclasses

+++ src/Logic_libary/Entity_Model.cpp
@@ -6,25 +6,25 @@ namespace logic {
 
     }
 
-    void Entity_Model::setX(double x) {m_x = x;}
+    void Entity_Model::setX(float x) {m_x = x;}
 
-    void Entity_Model::setY(double y) {m_y = y;}
+    void Entity_Model::setY(float y) {m_y = y;}
 
-    void Entity_Model::setVelocityX(double velocityX) {m_velocityX = velocityX;}
+    void Entity_Model::setVelocityX(float velocityX) {m_velocityX = velocityX;}
 
-    void Entity_Model::setVelocityY(double velocityY) {m_velocityY = velocityY;}
+    void Entity_Model::setVelocityY(float velocityY) {m_velocityY = velocityY;}
 
     void Entity_Model::setLenght(double lenght) {m_lenght = lenght;}
 
     void Entity_Model::setWidth(double width) {m_width = width;}
 
-    double Entity_Model::getX() const {return m_x;}
+    float Entity_Model::getX() const {return m_x;}
 
-    double Entity_Model::getY() const {return m_y;}
+    float Entity_Model::getY() const {return m_y;}
 
-    double Entity_Model::getVelocityX() const {return m_velocityX;}
+    float Entity_Model::getVelocityX() const {return m_velocityX;}
 
-    double Entity_Model::getVelocityY() const {return m_velocityY;}
+    float Entity_Model::getVelocityY() const {return m_velocityY;}
 
     double Entity_Model::getLenght() const {return m_lenght;}
